ECM-199: Suplier Simulator Backend Code refactor
